API: Move the router into its own file.
